refactor(app): move chat route to /session/$sessionId
Renames the chat session route from /chat/$sessionId to /session/$sessionId. Deletes the dead /chat index redirect (nothing navigated to the bare segment). Regenerates routeTree.gen.ts via the app build. Router-only: no backwards-compat redirect and no Chat* component rename.
