--- apps/app/src/routes/chat/index.tsx
@@ -1,10 +1,0 @@
-import { createFileRoute, Navigate } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/chat/")({
-  component: Component,
-});
-
-function Component() {
-  // Redirect to home page - sessions now require sessionId in URL
-  return <Navigate to="/" />;
-}
