+++ apps/app/src/routes/session/$sessionId.tsx
@@ -3,7 +3,7 @@ import { Button } from "@vibest/ui/components/button";
 
 import { Chat } from "@/components/chat";
 
-export const Route = createFileRoute("/chat/$sessionId")({
+export const Route = createFileRoute("/session/$sessionId")({
   component: Component,
 });
 
@@ -18,7 +18,7 @@ function Component() {
       const { sessionId: newSessionId } = await orpcQueryUtils.session.create.call({
         harnessAgentId: "claude-code",
       });
-      navigate({ to: "/chat/$sessionId", params: { sessionId: newSessionId } });
+      navigate({ to: "/session/$sessionId", params: { sessionId: newSessionId } });
     } catch (error) {
       console.error("Failed to start a new session", error);
     }
